requirements.txt
asgiref==3.8.1
black==24.10.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
Django==5.1.2
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.16.1
future==1.0.0
gevent==24.10.2
greenlet==3.1.1
idna==3.10
inflection==0.5.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
requests-file==2.1.0
sqlparse==0.5.1
tldextract==5.1.2
uritemplate==4.1.1
urllib3==2.2.3
zope.event==5.0
zope.interface==7.1.0
django-filter==24.3
Pillow==11.0.0
cloudinary==1.41.0 
django-cloudinary-storage==0.3.0 
six==1.16.0
gunicorn==23.0.0
python-telegram-bot==22.0
cryptography==46.0.1
django-redis==6.0.0
redis==6.4.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
